+++ BUILD.bazel
@@ -284,6 +284,7 @@ filegroup(
             "aten/src/ATen/cuda/detail/*.cpp",
             "aten/src/ATen/cudnn/*.cpp",
             "aten/src/ATen/native/cuda/*.cpp",
+            "aten/src/ATen/native/cuda/linalg/*.cpp",
             "aten/src/ATen/native/cudnn/*.cpp",
             "aten/src/ATen/native/miopen/*.cpp",
             "aten/src/ATen/native/sparse/cuda/*.cpp",
