+++ aten/src/ATen/native/cuda/linalg/BatchLinearAlgebra.cpp
@@ -12,7 +12,7 @@
 #include <ATen/native/Resize.h>
 #include <ATen/native/LinearAlgebra.h>
 #include <ATen/native/BatchLinearAlgebra.h>
-#include <ATen/native/cuda/BatchLinearAlgebraLib.h>
+#include <ATen/native/cuda/linalg/BatchLinearAlgebraLib.h>
 #include <ATen/native/cpu/zmath.h>
 
 #if AT_MAGMA_ENABLED()
