+++ aten/src/ATen/native/cuda/linalg/BatchLinearAlgebraLib.cpp
@@ -4,7 +4,6 @@
 #include <ATen/Dispatch.h>
 #include <ATen/NativeFunctions.h>
 #include <ATen/cuda/PinnedMemoryAllocator.h>
-#include <ATen/cuda/CUDASolver.h>
 #include <ATen/cuda/CUDABlas.h>
 #include <ATen/cuda/CUDAEvent.h>
 #include <c10/cuda/CUDAStream.h>
@@ -12,7 +11,8 @@
 
 #include <ATen/native/LinearAlgebraUtils.h>
 #include <ATen/native/cuda/MiscUtils.h>
-#include <ATen/native/cuda/BatchLinearAlgebraLib.h>
+#include <ATen/native/cuda/linalg/CUDASolver.h>
+#include <ATen/native/cuda/linalg/BatchLinearAlgebraLib.h>
 
 #ifndef AT_PER_OPERATOR_HEADERS
 #include <ATen/Functions.h>
