+++ aten/src/ATen/native/cuda/linalg/CUDASolver.cpp
@@ -1,6 +1,6 @@
 #include <ATen/Context.h>
 #include <ATen/NativeFunctions.h>
-#include <ATen/cuda/CUDASolver.h>
+#include <ATen/native/cuda/linalg/CUDASolver.h>
 #include <c10/cuda/CUDACachingAllocator.h>
 #include <c10/macros/Export.h>
